+++ dotnet/Razor/Scripts/hooks.ts
@@ -1,30 +1,40 @@
 // import cookie from 'cookie'
 // import { v4 as uuid } from '@lukeed/uuid'
 import type { Handle } from '@sveltejs/kit'
+import fs from 'node:fs'
+
 //import { RequestEvent } from '@sveltejs/kit/types/internal'
 
-export const handle: Handle = async ({ event, resolve }
-	// : { event:RequestEvent<Record<string,string>>, resolve:(RequestEvent<Record<string,string>>, ResolveOptions) => MaybePromise<Response>}
-	)=> {
+const [_beforeHead, _afterHead, _afterBody] = fs.readFileSync('./Scripts/app-page.html', {encoding: 'utf-8'}).split('%section%')
 
-	//console.assert(request, `hook handle req: ${JSON.stringify(request)}`)
+export const handle: Handle = async ({ event, resolve })=> {
 
-	// const cookies = cookie.parse(request.headers.cookie || '')
-	// request.locals.userid = cookies.userid || uuid()
+	const response = await resolve(event)
+
+	if (!event.request.headers.has('x-component')) {
+
+		const [headText, bodyText] = (await response.text()).split('%section%')
+
+		const finalResponse = new Response([
+			_beforeHead,
+			headText, 
+			_afterHead,
+			bodyText,
+			_afterBody
+		].join('\r\n'), { 
+			headers: response.headers,
+			status: response.status,
+			statusText: response.statusText
+		})
+		return finalResponse
+	}
 	
-	// TODO https://github.com/sveltejs/kit/issues/1046
-	// if (request.query.has('_method')) {
-	// 	request.method = request.query.get('_method').toUpperCase()
-	// }
-	
-	const response:Response = await resolve(event)
-	// console.warn('hook', response)
+	const sanitizedResponse = (await response.text()).replace('%section%','')
 
-	// if (!cookies.userid) {
-	// 	if this is the first time the user has visited this app,
-	// 	set a cookie so that we recognise them when they return
-	// 	response.headers['set-cookie'] = `userid=${request.locals.userid}; Path=/; HttpOnly`
-	// }
-
-	return response;
+	return new Response(sanitizedResponse, {			
+		headers: response.headers,
+		status: response.status,
+		statusText: response.statusText
+	})
 };
+