+++ dotnet/Razor/Scripts/routes/joke.ts
@@ -1,0 +1,4 @@
+import type { JSONValue, MaybePromise } from '@sveltejs/kit/types/private';
+import {getJoke} from '../lib/jokeService'
+
+export const get:()=> MaybePromise<{body?: JSONValue | Uint8Array | unknown;}> = async () => await getJoke()